bugfix 任务的已提交的任务，engineid为空

+++ service/src/main/java/com/dtstack/rdos/engine/service/zk/task/TaskStatusListener.java
@@ -164,12 +164,14 @@ public class TaskStatusListener implements Runnable{
 	private void dealStreamJob(String taskId, String engineTypeName, String zkTaskId, int computeType, Integer oldStatus) throws Exception {
         RdosEngineStreamJob rdosTask = rdosStreamTaskDAO.getRdosTaskByTaskId(taskId);
 
+        boolean normalJob = false;
         if(rdosTask != null){
             String engineTaskId = rdosTask.getEngineTaskId();
             String appId = rdosTask.getApplicationId();
             JobIdentifier jobIdentifier = JobIdentifier.createInstance(engineTaskId, appId, taskId);
 
             if(StringUtils.isNotBlank(engineTaskId)){
+                normalJob = true;
                 String pluginInfoStr = "";
                 if(rdosTask.getPluginInfoId() > 0 ){
                     pluginInfoStr = pluginInfoDao.getPluginInfo(rdosTask.getPluginInfoId());
@@ -197,7 +199,9 @@ public class TaskStatusListener implements Runnable{
                     addFailedJob(failedTaskInfo);
                 }
             }
-        } else {
+        }
+
+        if (!normalJob){
             zkLocalCache.updateLocalMemTaskStatus(zkTaskId, RdosTaskStatus.FAILED.getStatus());
             rdosEngineJobCacheDao.deleteJob(taskId);
         }
@@ -206,11 +210,13 @@ public class TaskStatusListener implements Runnable{
     private void dealBatchJob(String taskId, String engineTypeName, String zkTaskId, int computeType, Integer oldStatus) throws Exception {
         RdosEngineBatchJob rdosBatchJob  = rdosBatchEngineJobDAO.getRdosTaskByTaskId(taskId);
 
+        boolean normalJob = false;
         if(rdosBatchJob != null){
             String engineTaskId = rdosBatchJob.getEngineJobId();
             JobIdentifier jobIdentifier = JobIdentifier.createInstance(engineTaskId, null, null);
 
             if(StringUtils.isNotBlank(engineTaskId)){
+                normalJob = true;
                 String pluginInfoStr = "";
                 if(rdosBatchJob.getPluginInfoId() > 0 ){
                     pluginInfoStr = pluginInfoDao.getPluginInfo(rdosBatchJob.getPluginInfoId());
@@ -238,7 +244,9 @@ public class TaskStatusListener implements Runnable{
                     addFailedJob(failedTaskInfo);
                 }
             }
-        } else {
+        }
+
+        if (!normalJob){
             zkLocalCache.updateLocalMemTaskStatus(zkTaskId, RdosTaskStatus.FAILED.getStatus());
             rdosEngineJobCacheDao.deleteJob(taskId);
         }
